% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHOPStudy-package.R
\docType{package}
\name{CHOPStudy-package}
\alias{CHOPStudy}
\alias{CHOPStudy-package}
\title{CHOPStudy: Research Compendium for the CHOP Study}
\description{
This package is a research compendium for a study examining the prevalence of bandemia in canine patients undergoing chemotherapy treatment. The research team is led by Dr. Alison Masyr (PI) at Michigan State University.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sjpierce/CHOPStudy}
}

}
\author{
\strong{Maintainer}: Steven J. Pierce \email{pierces1@msu.edu} (\href{https://orcid.org/0000-0002-0679-3019}{ORCID}) [copyright holder]

}
\keyword{internal}
